// apps/backend/src/adapters/erp/logo/logo.types.ts
export interface LogoItemRow {
  LOGICALREF?: number;
  CODE?: string;
  NAME?: string;
  ONHAND?: number;
  SALESPRICE?: number;
  BARCODE?: string;
}

export interface LogoCurrentRow {
  LOGICALREF?: number;
  CODE?: string;
  TITLE?: string;
}

export interface LogoSalesInvoiceRow {
  LOGICALREF?: number;
  FICHENO?: string;
  NUMBER?: string;
}